remove kube.go and dependencies
