+++ callback_test.go
@@ -370,25 +370,3 @@ func TestArgoApplication(t *testing.T) {
 	// print time taken
 	fmt.Printf("Time taken for json.Unmarshal(): %v\n", time.Since(begin))
 }
-
-func TestEmptyJson(t *testing.T) {
-	input := []byte(`{
-
-	}`)
-	callbacks := []*MatchCallback{}
-	opt := MatchOptions{
-		IgnoreCase:       false,
-		QuitIfNoCallback: false,
-	}
-	res := MatchResult{
-		Count: 0,
-	}
-	err := Match(input, &opt, &res, callbacks...)
-	if err != nil {
-		t.Fatal(err)
-	}
-	// print count
-	if res.Count != 0 {
-		t.Fatal("expected count to be 0")
-	}
-}
